✨ 페이지 래퍼 컴포넌트 구조 개선 및 CenteredPageWrapper 추가
